Challenge48h/2021/Pomona-Passion_froid-Mediatheque/requirements.txt
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
SQLAlchemy==1.3.16
PyMySQL==0.9.3
Werkzeug==1.0.1
requests==2.23.0
Flask-Login==0.5.0
Pillow==8.1.0
